api/requirements.txt
Django==3.2
django-cockroachdb==3.2.*
psycopg2
django-environ
djangorestframework
